refactor(scaffold): extract app init template to standalone SDK v2 Python file
Move the inline `r#"..."#` Python template from `scaffold_python_app()` in
src/cli.rs to sdk/python/plexi_sdk/templates/app_init.py and load it via
`include_str!`. Rewrite the template from low-level primitives (ctx.rect,
ctx.text, Button) to SDK v2 declarative components (Column, AppBar, Card,
KeyRow, FooterKeys, etc.).

Closes #1271
