+++ sdk/python/plexi_sdk/templates/app_init.py
@@ -1,0 +1,60 @@
+#!/usr/bin/env python3
+"""
+__DISPLAY_NAME__ — generated by `plexi app init`.
+
+Delete what you don't need.
+"""
+from plexi_sdk import App, RenderContext
+from plexi_sdk.ui import (
+    AppBar,
+    Card,
+    Column,
+    FooterKeys,
+    KeyRow,
+    Label,
+    Section,
+    Spacer,
+)
+
+
+class __CLASS_NAME__(App):
+    async def on_init(self, ctx: RenderContext) -> None:
+        self.click_count = 0
+        ctx.status_summary("Ready")
+        self.emit.info("__DISPLAY_NAME__ initialized")
+
+    def on_render(self, ctx: RenderContext) -> None:
+        info_rows = [
+            f"app_id: {self.app_id}",
+            f"workspace: {ctx.workspace_root or '(none)'}",
+            f"dimensions: {ctx.w:.0f} x {ctx.h:.0f}",
+            f"capabilities: {', '.join(ctx.capabilities) or 'none'}",
+        ]
+
+        ctx.render(Column([
+            AppBar(title="__DISPLAY_NAME__", subtitle="Edit main.py to build your app"),
+            Section("App Info"),
+            Card([Label(line, tone="caption") for line in info_rows]),
+            Section("Keyboard Shortcuts"),
+            Card([
+                KeyRow("q", "quit"),
+                KeyRow("+", "increment counter"),
+            ]),
+            Label(f"Clicks: {self.click_count}", tone="body"),
+            Spacer(grow=True),
+            FooterKeys([
+                ("q", "quit"),
+                ("+", "increment"),
+            ]),
+        ]))
+
+    def on_key(self, ctx: RenderContext, key: str, mods: dict) -> None:
+        if key == "+" or (key == "=" and mods.get("shift")):
+            self.click_count += 1
+            ctx.status_summary(f"Clicked {self.click_count} time(s)")
+
+    def on_click(self, ctx: RenderContext, x: float, y: float, button: str) -> None:
+        pass
+
+
+__CLASS_NAME__().run()
