+++ src/cli.rs
@@ -565,108 +565,7 @@ fn scaffold_python_app(app_dir: &std::path::Path, name: &str) -> io::Result<()> 
     // do NOT copy plexi_sdk.py alongside (the package uses relative imports
     // that break when imported as a flat single file).
     // __CLASS_NAME__ and __DISPLAY_NAME__ are substituted below.
-    let template = r#"#!/usr/bin/env python3
-"""
-__DISPLAY_NAME__ — generated by `plexi app init`.
-
-Delete what you don't need.
-"""
-from plexi_sdk import (
-    App, RenderContext,
-    BG, SURFACE, FG, ACCENT, MUTED, HIGHLIGHT,
-    BODY, CAPTION, HEADING, HINT,
-    GREEN,
-    PAD, PAD_TIGHT,
-    PRIORITY_NORMAL,
-)
-from plexi_sdk.widgets import Button, ButtonStyle
-
-
-class __CLASS_NAME__(App):
-    async def on_init(self, ctx: RenderContext) -> None:
-        self.click_count = 0
-        ctx.status_summary("Ready")
-        self.emit.info("__DISPLAY_NAME__ initialized")
-
-        self._btn = Button(
-            "increment", x=PAD, y=0, w=130, h=30, label="[ Click me ]",
-            style=ButtonStyle(
-                fill=SURFACE, hover_fill=HIGHLIGHT, active_fill=HIGHLIGHT,
-                text_color=ACCENT, font_size=CAPTION, radius=3.0,
-            ),
-        )
-
-    def on_render(self, ctx: RenderContext) -> None:
-        # ctx.w / ctx.h — current pane dimensions; ctx.elapsed — seconds since last frame
-        ctx.clear(BG)
-
-        # ── Header ───────────────────────────────────────────────────────────────
-        # ctx.row positions children side-by-side with host-side flexbox.
-        # badge_child / text_child / key_chip_child build the children list.
-        ctx.row(
-            x=PAD, y=PAD,
-            children=[
-                ctx.text_child("__DISPLAY_NAME__", size=HEADING, color=ACCENT,
-                               bold=True, monospace=True),
-                ctx.badge_child("v0.1.0", fill=SURFACE, fg=MUTED,
-                                font_size=HINT, radius=3.0),
-            ],
-            gap=8.0,
-        )
-
-        subtitle_y = PAD + HEADING + PAD_TIGHT
-        ctx.text(PAD, subtitle_y, "Edit main.py to build your app",
-                 size=CAPTION, color=MUTED)
-
-        # ── Info table ────────────────────────────────────────────────────────────
-        # Two-column table: keys at a fixed indent, values at KEY_COL offset.
-        # Use ctx.row() for flex items (badges, chips); use fixed x for aligned
-        # table columns where values must line up vertically.
-        rows = [
-            ("app_id",     self.app_id),
-            ("workspace",  ctx.workspace_root or "(none)"),
-            ("dimensions", f"{ctx.w:.0f} × {ctx.h:.0f}"),
-            ("caps",       ", ".join(ctx.capabilities) or "none"),
-        ]
-        ROW_H   = 30
-        KEY_COL = 100
-        table_y = subtitle_y + CAPTION + PAD
-        table_h = ROW_H * len(rows)
-
-        ctx.rect(PAD, table_y, ctx.w - PAD * 2, table_h, fill=SURFACE, radius=6.0)
-
-        for i, (key, value) in enumerate(rows):
-            row_y = table_y + i * ROW_H
-            if i > 0:
-                ctx.rect(PAD, row_y, ctx.w - PAD * 2, 1, fill=BG)
-            text_y = row_y + (ROW_H - CAPTION) / 2
-            ctx.text(PAD + 10,            text_y, key,   size=CAPTION, color=GREEN, monospace=True)
-            ctx.text(PAD + 10 + KEY_COL, text_y, value, size=CAPTION, color=FG,    monospace=True,
-                     max_width=ctx.w - PAD * 2 - 10 - KEY_COL, elide=True)
-
-        # ── Button + state ────────────────────────────────────────────────────────
-        self._btn.y = table_y + table_h + PAD
-        if self._btn.render(ctx):
-            self.click_count += 1
-            ctx.status_summary(f"Clicked {self.click_count} time(s)")
-            # ctx.notify("__DISPLAY_NAME__", f"Click #{self.click_count}", priority=PRIORITY_NORMAL)
-
-        ctx.text(PAD, self._btn.y + self._btn.h + PAD_TIGHT,
-                 f"Clicks: {self.click_count}", size=BODY, color=ACCENT)
-
-    def on_key(self, ctx: RenderContext, key: str, mods: dict) -> None:
-        # key: "a"–"z", "0"–"9", "up"/"down"/"left"/"right", "return", "escape", etc.
-        # mods: {"shift": bool, "ctrl": bool, "alt": bool, "meta": bool}
-        if key == "q":
-            pass
-
-    def on_click(self, ctx: RenderContext, x: float, y: float, button: str) -> None:
-        # button: "primary" | "secondary" | "middle"
-        pass
-
-
-__CLASS_NAME__().run()
-"#;
+    let template = include_str!("../sdk/python/plexi_sdk/templates/app_init.py");
     let main_py = template
         .replace("__CLASS_NAME__", &to_struct_name(name))
         .replace("__DISPLAY_NAME__", &to_title_case(name));
